show friends of friends in trip and group feed fix logout fix refresh data when back to feed screen
